refactor: split builders into ProcessApiBuilder and SharedTypesBuilder
Each language now has two dedicated classes with a single responsibility:
- KotlinProcessApiBuilder / JavaProcessApiBuilder — generate the per-process API file
- KotlinSharedTypesBuilder / JavaSharedTypesBuilder — generate the 5 shared type files

CodeGenerationAdapter wires both and combines their results. Tests and
smoke tests updated accordingly; `!!` replaced with requireNotNull throughout.
