+++ bpmn-to-code-core/src/main/kotlin/io/github/emaarco/bpmn/adapter/outbound/codegen/CodeGenerationAdapter.kt
@@ -1,31 +1,46 @@
 package io.github.emaarco.bpmn.adapter.outbound.codegen
 
-import io.github.emaarco.bpmn.adapter.outbound.codegen.builder.JavaApiBuilder
-import io.github.emaarco.bpmn.adapter.outbound.codegen.builder.KotlinApiBuilder
+import io.github.emaarco.bpmn.adapter.outbound.codegen.builder.JavaProcessApiBuilder
+import io.github.emaarco.bpmn.adapter.outbound.codegen.builder.JavaSharedTypesBuilder
+import io.github.emaarco.bpmn.adapter.outbound.codegen.builder.KotlinProcessApiBuilder
+import io.github.emaarco.bpmn.adapter.outbound.codegen.builder.KotlinSharedTypesBuilder
 import io.github.emaarco.bpmn.application.port.outbound.GenerateApiCodePort
 import io.github.emaarco.bpmn.domain.BpmnModelApi
 import io.github.emaarco.bpmn.domain.GeneratedApiFile
 import io.github.emaarco.bpmn.domain.shared.OutputLanguage
 
 class CodeGenerationAdapter(
-    private val builder: Map<OutputLanguage, AbstractApiBuilder<*>> = Companion.builder
+    private val processApiBuilders: Map<OutputLanguage, AbstractProcessApiBuilder<*>> = Companion.processApiBuilders,
+    private val sharedTypesBuilders: Map<OutputLanguage, AbstractSharedTypesBuilder> = Companion.sharedTypesBuilders,
 ) : GenerateApiCodePort {
 
     override fun generateCode(modelApi: BpmnModelApi): List<GeneratedApiFile> {
-        val outputLanguage = modelApi.outputLanguage
-        val builder = builder[outputLanguage] ?: throw IllegalArgumentException("$outputLanguage is not supported")
-        return builder.buildApiFile(modelApi)
+        val language = modelApi.outputLanguage
+        val processApiBuilder = processApiBuilders[language] ?: throw IllegalArgumentException("$language is not supported")
+        val sharedTypesBuilder = sharedTypesBuilders[language] ?: throw IllegalArgumentException("$language is not supported")
+        val processFile = processApiBuilder.buildApiFile(modelApi)
+        val typeFiles = sharedTypesBuilder.buildTypeFiles(modelApi.packagePath, language)
+        return listOf(processFile) + typeFiles
     }
 
-    abstract class AbstractApiBuilder<MainBuilder : Any> {
-        abstract fun buildApiFile(modelApi: BpmnModelApi): List<GeneratedApiFile>
+    abstract class AbstractProcessApiBuilder<MainBuilder : Any> {
+        abstract fun buildApiFile(modelApi: BpmnModelApi): GeneratedApiFile
+        protected val autoGenComment = "Generated by bpmn-to-code"
+    }
+
+    abstract class AbstractSharedTypesBuilder {
+        abstract fun buildTypeFiles(packagePath: String, language: OutputLanguage): List<GeneratedApiFile>
         protected val autoGenComment = "Generated by bpmn-to-code"
     }
 
     companion object {
-        val builder = mapOf(
-            OutputLanguage.KOTLIN to KotlinApiBuilder(),
-            OutputLanguage.JAVA to JavaApiBuilder()
+        val processApiBuilders = mapOf(
+            OutputLanguage.KOTLIN to KotlinProcessApiBuilder(),
+            OutputLanguage.JAVA to JavaProcessApiBuilder()
+        )
+        val sharedTypesBuilders = mapOf(
+            OutputLanguage.KOTLIN to KotlinSharedTypesBuilder(),
+            OutputLanguage.JAVA to JavaSharedTypesBuilder()
         )
     }
