+++ bpmn-to-code-core/src/test/kotlin/io/github/emaarco/bpmn/adapter/outbound/codegen/builder/NewsletterFlowNodes.kt
@@ -2,65 +2,12 @@ package io.github.emaarco.bpmn.adapter.outbound.codegen.builder
 
 import io.github.emaarco.bpmn.domain.shared.BpmnElementType
 import io.github.emaarco.bpmn.domain.shared.CallActivityDefinition
-import io.github.emaarco.bpmn.domain.shared.EscalationDefinition
 import io.github.emaarco.bpmn.domain.shared.FlowNodeDefinition
 import io.github.emaarco.bpmn.domain.shared.FlowNodeProperties
 import io.github.emaarco.bpmn.domain.shared.ServiceTaskDefinition
 import io.github.emaarco.bpmn.domain.shared.ServiceTaskDefinition.Companion.IMPL_VALUE_KEY
 import io.github.emaarco.bpmn.domain.shared.TimerDefinition
 import io.github.emaarco.bpmn.domain.shared.VariableDefinition
-import io.github.emaarco.bpmn.domain.testBpmnModelApi
-import io.github.emaarco.bpmn.domain.testNewsletterBpmnModel
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Test
-import java.io.File
-
-class KotlinApiBuilderTest {
-
-    private val underTest = KotlinApiBuilder()
-
-    @Test
-    fun `buildApiFile generates correct API file content`() {
-
-        // given: a BPMN model with custom service task implementations
-        val modelApi = testBpmnModelApi(
-            packagePath = "de.emaarco.example",
-            model = testNewsletterBpmnModel(
-                flowNodes = buildNewsletterFlowNodes(
-                    confirmationMailImpl = "#{newsletterSendConfirmationMail}",
-                    welcomeMailImpl = "\${newsletterSendWelcomeMail}",
-                    registrationCompletedImpl = "newsletter.registrationCompleted",
-                    extraVariables = listOf(VariableDefinition("testVariable")),
-                ),
-                escalations = listOf(EscalationDefinition("EndEvent_RegistrationNotPossible", "Escalation_RegistrationFailed", "100"))
-            )
-        )
-
-        // when: we build the API file
-        val results = underTest.buildApiFile(modelApi)
-
-        // then: 6 files returned — 1 model + 5 shared types
-        assertThat(results).hasSize(6)
-        val typeFiles = results.filter { it.packagePath == "de.emaarco.example.types" }
-        assertThat(typeFiles.map { it.fileName }).containsExactlyInAnyOrder(
-            "BpmnTimer.kt", "BpmnError.kt", "BpmnEscalation.kt", "BpmnFlow.kt", "BpmnRelations.kt"
-        )
-
-        val modelFile = results.first { it.fileName == "${modelApi.fileName()}.kt" }
-        val expectedFile = File(javaClass.getResource("/api/NewsletterSubscriptionProcessApiKotlin.txt").toURI())
-        assertThat(modelFile.content).isEqualToIgnoringWhitespace(expectedFile.readText())
-        assertThat(modelFile.packagePath).isEqualTo("de.emaarco.example")
-
-        // and: each type file matches its expected fixture
-        typeFiles.forEach { typeFile ->
-            val fixtureName = typeFile.fileName.replace(".kt", "Kotlin.txt")
-            val fixtureResource = javaClass.getResource("/api/types/$fixtureName")
-                ?: error("Missing fixture: /api/types/$fixtureName")
-            assertThat(typeFile.content).isEqualToIgnoringWhitespace(File(fixtureResource.toURI()).readText())
-        }
-    }
-
-}
 
 internal fun buildNewsletterFlowNodes(
     confirmationMailImpl: String,
