+++ bpmn-to-code-gradle/src/test/kotlin/io/github/emaarco/bpmn/adapter/GradlePluginSmokeTest.kt
@@ -25,7 +25,7 @@ class GradlePluginSmokeTest {
     ) {
         // Copy BPMN file into the temp project
         val resourcesDir = File(projectDir, "src/main/resources").also { it.mkdirs() }
-        val bpmnStream = javaClass.classLoader.getResourceAsStream("bpmn/$bpmnFile")!!
+        val bpmnStream = requireNotNull(javaClass.classLoader.getResourceAsStream("bpmn/$bpmnFile"))
         File(resourcesDir, bpmnFile).writeBytes(bpmnStream.readBytes())
 
         // Write settings.gradle
@@ -62,17 +62,13 @@ class GradlePluginSmokeTest {
         // Verify output files were generated
         val packageDir = File(projectDir, "build/generated/io/github/emaarco/smoketest")
         assertThat(packageDir).isDirectory()
-        val generatedFiles = packageDir.listFiles()!!
+        val generatedFiles = requireNotNull(packageDir.listFiles())
         assertThat(generatedFiles).isNotEmpty()
 
         val expectedExt = if (language == "KOTLIN") ".kt" else ".java"
         val modelFiles = generatedFiles.filter { it.isFile }
-        assertThat(modelFiles).allSatisfy { file ->
-            assertThat(file.name).endsWith(expectedExt)
-        }
+        assertThat(modelFiles).allSatisfy { file -> assertThat(file.name).endsWith(expectedExt) }
         val typesDir = generatedFiles.first { it.isDirectory && it.name == "types" }
-        assertThat(typesDir.listFiles()!!).allSatisfy { file ->
-            assertThat(file.name).endsWith(expectedExt)
-        }
+        assertThat(requireNotNull(typesDir.listFiles())).allSatisfy { file -> assertThat(file.name).endsWith(expectedExt) }
     }
 }
